Move eval logic to fasteval project
