+++ pyproject.toml
@@ -16,6 +16,7 @@ requires-python = ">=3.10"
 dependencies = [
   # TODO: To use `uv`, you need to do `uv pip install --prerelease=allow .` so it recognizes pre-release AutoGluon
   "autogluon>=1.4.1b20250910,<1.5",  # TODO: Remove after moving `benchmark` code elsewhere
+  "fasteval @ file:///${PROJECT_ROOT}/fasteval",
   "openml>=0.14.1",  # consider making optional
   "pyyaml",
   "pytest",
