+++ tabarena/nips2025_utils/scripts/run_amlb2025.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import pandas as pd
 
 from autogluon.tabular import TabularDataset
-from tabarena.tabarena.tabarena import TabArena
+from fasteval.tabarena import TabArena
 
 
 if __name__ == '__main__':
