+++ tabarena/paper/paper_runner_tabarena.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import pandas as pd
 
 from .paper_runner import PaperRun
-from tabarena.tabarena.tabarena import TabArena
+from fasteval.tabarena import TabArena
 
 
 class PaperRunTabArena(PaperRun):
