+++ tst/test_elo_utils.py
@@ -1,5 +1,5 @@
 import pytest
-from tabarena.tabarena.elo_utils import EloHelper
+from fasteval.elo_utils import EloHelper
 import pandas as pd
